# Conflicts:
#   frontend/src/components/welcome/WelcomeHeader.tsx

+++ frontend/src/components/welcome/OAuth2RedirectHandler.tsx
@@ -2,11 +2,17 @@ import { useEffect } from "react";
 
 export const OAuth2RedirectHandler = () => {
   useEffect(() => {
-    localStorage.clear();
-    localStorage.setItem(
-      "token",
-      window.location.search.substring(7) as string
-    );
+    function searchParam(key: string) {
+      return new URLSearchParams(window.location.search).get(key);
+    }
+    const token = searchParam("token");
+    const error = searchParam("error");
+    if (token) {
+      localStorage.clear();
+      localStorage.setItem("token", token);
+    } else if (error) {
+      alert(error);
+    }
     window.location.replace("/");
   }, []);
   return <div></div>;
